Changes to FEApp aztecoo epetra phalanx trilinoscouplings xpetra so that almost all tests possibly affected by Epetra 64 bit changes pass.  Most of the issues are related to the ambiguity created by int/long long overloading and the passed in type something else (like unsigned int).  Other issue fixed here is to use int based maps unless there is a reason to use long long map.

+++ demos/FEApp/src/FEApp_CZeroDiscretization.cpp
@@ -52,7 +52,7 @@ FEApp::CZeroDiscretization::CZeroDiscretization(
   neq(num_equations)
 {
   // Distribute the elements equally among processors
-  elem_map = Teuchos::rcp(new Epetra_Map(x.size()-1, 0, *comm));
+  elem_map = Teuchos::rcp(new Epetra_Map(static_cast<int>(x.size()-1), 0, *comm));
   numMyElements = elem_map->NumMyElements();
 
   Teuchos::RCP<AbstractElement> base_element =
